6351 | 12 files | M DumpSource2/schemas/client/CDOTA_Ability_Broodmother_StickySnare.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Special_Mars_Spear_Burning_Trail_Thinker.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Viper_Nose_Dive.h,  M DumpSource2/schemas/server/CDOTA_Ability_Broodmother_StickySnare.h,  M DumpSource2/schemas/server/CDOTA_BaseNPC_Hero.h,  M DumpSource2/schemas/server/CDOTA_Modifier_Special_Mars_Spear_Burning_Trail_Thinker.h,  M DumpSource2/schemas/server/CDOTA_Modifier_Viper_Nose_Dive.h,  M game/bin/built_from_cl.txt,  M game/dota/bin/linuxsteamrt64/libclient_strings.txt,  M game/dota/bin/linuxsteamrt64/libserver_strings.txt,  M game/dota/pak01_dir.txt,  M game/dota/steam.inf
https://steamdb.info/patchnotes/17684232/

+++ DumpSource2/schemas/client/CDOTA_Modifier_Viper_Nose_Dive.h
@@ -4,8 +4,8 @@ class CDOTA_Modifier_Viper_Nose_Dive
 	Vector m_vStart;
 	float32 m_flZDelta;
 	float32 radius;
-	int32 corrosive_radius;
+	float32 corrosive_radius;
 	float32 effect_duration;
-	int32 dive_speed;
+	float32 dive_speed;
 	float32 start_height;
 };
