+++ DumpSource2/schemas/server/CDOTA_Modifier_Special_Mars_Spear_Burning_Trail_Thinker.h
@@ -1,7 +1,7 @@
 class CDOTA_Modifier_Special_Mars_Spear_Burning_Trail_Thinker
 {
 	Vector m_vPathDir;
-	int32 shard_trail_radius;
+	float32 shard_trail_radius;
 	float32 shard_debuff_linger_duration;
 	float32 shard_interval;
 };
